+++ src/constants/shine.constant.ts
@@ -1,0 +1,13 @@
+interface ShineConfig {
+  width: string;
+  top: string;
+  right: string;
+}
+
+export const SHINE_CONFIGS: ShineConfig[] = [
+  { width: '111px', top: '180px', right: '3px' },
+  { width: '57px', top: '90px', right: '-170px' },
+  { width: '57px', top: '50px', right: '10px' },
+  { width: '101px', top: '190px', right: '0px' },
+  { width: '65px', top: '20px', right: '150px' },
+];
